requirements.txt
opencv-python
torch==2.2.2
pandas==2.2.2
numpy>=1.25.0,<2.0.0
requests==2.32.2
ultralytics==8.1.34
supervision==0.18.0
azure-storage-blob==12.19.1
shapely>=2.0.1,<3.0.0
librosa>=0.10.0,<1.0.0
soundfile>=0.12.1,<1.0.0
pyodbc>=4.0.34,<5.0.0